Rename SkPaintImageFilter to SkShaderImageFilter
Bug: skia:11230
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/371965
