+++ gn/effects_imagefilters.gni
@@ -26,10 +26,9 @@ skia_effects_imagefilter_sources = [
   "$_src/effects/imagefilters/SkMergeImageFilter.cpp",
   "$_src/effects/imagefilters/SkMorphologyImageFilter.cpp",
   "$_src/effects/imagefilters/SkOffsetImageFilter.cpp",
-  "$_src/effects/imagefilters/SkPaintImageFilter.cpp",
-  "$_src/effects/imagefilters/SkPaintImageFilter.h",
   "$_src/effects/imagefilters/SkPictureImageFilter.cpp",
   "$_src/effects/imagefilters/SkPictureImageFilter.h",
+  "$_src/effects/imagefilters/SkShaderImageFilter.cpp",
   "$_src/effects/imagefilters/SkTileImageFilter.cpp",
   "$_src/effects/imagefilters/SkTileImageFilter.h",
 ]
